# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/PostCountryDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class PostCountryDTO {

    private Long amountOfCountryToSave;
}